PHPStan level 0 fixes #1990
